+++ src/main/java/exter/foundry/gui/GuiMetalCaster.java
@@ -93,7 +93,7 @@ public class GuiMetalCaster extends GuiFoundry {
 		}
 
 		//Draw stored power bar.
-		int power = (int) (te_caster.getStoredFoundryEnergy() * POWER_HEIGHT / te_caster.getFoundryEnergyCapacity());
+		int power = te_caster.getStoredFoundryEnergy() * POWER_HEIGHT / te_caster.getFoundryEnergyCapacity();
 		if (power > 0) {
 			drawTexturedModalRect(window_x + POWER_X, window_y + POWER_Y + POWER_HEIGHT - power, POWER_OVERLAY_X, POWER_OVERLAY_Y + POWER_HEIGHT - power, POWER_WIDTH, power);
 		}
